+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+0.4.0
+-----
+ * Fix crash when setting main touchbar icon
+ * Fix crash when refreshing scrubber with no selected item
+ * Allocate image items instead of using paths
+ * Support for Apple template images
+ * Support for spacer elements
+ * Convenience function for finding bundled images
+ * Support setting label and button widths
+ * Support non-continous sliders
+ * Support changing button contents
+ * Support label for sliders
+ * Support adding swipe gestures to touch bar items
+ * Fix font in multi-line labels.
+ * Fix invalid items sent to button callbacks.
+
 0.3.0
 -----
  * Redesigned to track and release allocated memory.  Most leaks fixed.